some things
